+++ src/Netsphere.Database/Game/PlayerDenyEntity.cs
@@ -10,11 +10,11 @@ namespace Netsphere.Database.Game
         public long Id { get; set; }
 
         [Column]
-        public long PlayerId { get; set; }
+        public int PlayerId { get; set; }
         public PlayerEntity Player { get; set; }
 
         [Column]
-        public long DenyPlayerId { get; set; }
+        public int DenyPlayerId { get; set; }
         public PlayerEntity DenyPlayer { get; set; }
     }
 }
